Refactor entire project
The `server.py` file was getting to monolithic, the project was
refactored into a bunch of smaller files. In the process, I also
refactored a lot of logic without testing so a lot of things are
probably broken. I'm just going to throw out the old test suite and
just write an entirely new one.